Create forcesshdconf.yml
Another, clearer example of the basic idea behind Ansible is the following task, which allows us to periodically overwrite configuration discrepancies in the SSHD config. It may be that I need to switch to password-based login at some point and change the config, or a future upgrade of the SSHD config discards the previous changes. It may even be that I forget to undo the change. Whatever happens!